progress on login and register

+++ src/containers/register/Register.jsx
@@ -1,28 +1,52 @@
 import React, { useEffect, useState } from 'react';
+import { useHistory } from 'react-router-dom';
+
+// material
 import Box from '@mui/material/Box';
 import TextField from '@mui/material/TextField'
 import Button from '@mui/material/Button';
+import ToggleButton from '@mui/material/ToggleButton';
+import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
 
+// firebase
 import '../../firebase-config';
 import { getAuth, createUserWithEmailAndPassword } from 'firebase/auth'
-import { useHistory } from 'react-router-dom';
+import { getDatabase, ref, set } from "firebase/database";
 
+// toast
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
 const authentication = getAuth()
+const database = getDatabase()
 
 const Register = () => {
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('')
+    const [confirmPassword, setConfirmPassword] = useState('')
+    const [userRole, setUserRole] = useState('player')
     const history = useHistory();
 
     const handleRegister = () => {
-        createUserWithEmailAndPassword(authentication, email, password)
+
+        if (!password || !email || !confirmPassword) {
+            toast.error('Please fill out all of the fields.')
+        } else if (password !== confirmPassword) {
+            toast.error('Passwords do not match.')
+        } else {
+            createUserWithEmailAndPassword(authentication, email, password)
             .then((response) => {
                 console.log(response)
+                const userData = response?.user
                 sessionStorage.setItem('Auth Token', response._tokenResponse.refreshToken)
-                history.push('/')
+                
+                set(ref(database, 'users/' + userData.uid), {
+                    email: userData.email,
+                    role: userRole
+                });
+
+                history.push('/profile')
+                toast.success("You have registerd and logged in successfully.")
             })
             .catch((error) => {
                 console.log(error.code)
@@ -34,6 +58,7 @@ const Register = () => {
                     toast.error('An error has occured. Please try again.')
                 }
             })
+        }
     }
 
     useEffect(() => {
@@ -43,31 +68,51 @@ const Register = () => {
     return (
         <div className="flex-column justify-center align-center container">
             <h3>Register Form</h3>
-            <Box
-                component="form"
-                sx={{
-                    '& > :not(style)': { m: 1, width: '25ch' },
+            <div>I want to register as a:</div>
+            <ToggleButtonGroup
+                color="primary"
+                value={userRole}
+                sx={{height: 40, margin: '10px 5px 5px 0px'}}
+                exclusive
+                onChange={(e) => {
+                    setUserRole(e.target.value)
                 }}
-                noValidate
-                autoComplete="off"
-            >
+                >
+                <ToggleButton value="clubRep">Club Rep</ToggleButton>
+                <ToggleButton value="player">Player</ToggleButton>
+            </ToggleButtonGroup>
+            <div className="flex-column">
                 <TextField 
                     id="email" 
-                    label="Enter email" 
+                    label="Email*" 
                     variant="outlined" 
                     value={email}
+                    size="small"
+                    sx={{marginTop: '15px'}}
                     onChange={(e) => setEmail(e.target.value)}
                 />
                 <TextField 
                     id="password" 
                     type="password"
-                    label="Enter password" 
+                    label="Password*" 
                     variant="outlined" 
                     value={password}
+                    size="small"
+                    sx={{marginTop: '15px'}}
                     onChange={(e) => setPassword(e.target.value)}
                 />
-            </Box>
-            <Button variant="contained" sx={{margin: '0 !important'}} onClick={handleRegister}>Register</Button>
+                <TextField 
+                    id="confirmPassword" 
+                    type="password"
+                    label="Confirm password*" 
+                    variant="outlined" 
+                    value={confirmPassword}
+                    size="small"
+                    sx={{marginTop: '15px'}}
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                />
+            </div>
+            <Button variant="contained" sx={{margin: '10px !important'}} type="submit" onClick={handleRegister} disabled={!email || !password || !confirmPassword}>Register</Button>
             <ToastContainer autoClose={3000}/>
         </div>
     )
